adding sync version of filter functions
